+++ UserTools/PrintGenieEvent/PrintGenieEvent.cpp
@@ -19,15 +19,17 @@ bool PrintGenieEvent::Execute(){
 	// Get the genie event
 	
 //	intptr_t thegenieinfoptr;
-	GenieInfo* thegenieinfoptr;
-	int get_ok = m_data->Stores.at("GenieInfo")->Get("TheGenieInfoPtr",thegenieinfoptr);
+	//GenieInfo* thegenieinfoptr;
+	//int get_ok = m_data->Stores.at("GenieInfo")->Get("TheGenieInfoPtr",thegenieinfoptr);
+	GenieInfo thegenieinfo;
+	int get_ok = m_data->Stores.at("GenieInfo")->Get("GenieInfo",thegenieinfo);
 	
 	if(!get_ok){
 		Log("PrintGenieEvent Tool: Failed to retrieve TheGenieInfo rom GenieInfo BoostStore!",1,0);
 		return false;
 	}
 	//GenieInfo* thegenieinfop = reinterpret_cast<GenieInfo*>(thegenieinfoptr);
-	GenieInfo thegenieinfo = (*thegenieinfoptr);
+	//GenieInfo thegenieinfo = (*thegenieinfoptr);
 	
 	// Print the info
 	cout<<"This was a "<< thegenieinfo.procinfostring <<" (neut code "<<thegenieinfo.neutinteractioncode
