+++ configfiles/LoadGenieEvents/LoadGenieEventConfig
@@ -1,4 +1,4 @@
-verbosity 100
+verbosity 1
 FluxVersion 1  # use 0 to load genie files based on bnb_annie_0000.root etc files
                # use 1 to load files based on beammc_annie_0000.root etc files
 #FileDir /pnfs/annie/persistent/users/vfischer/genie/BNB_Water_10k_22-05-17/
